requirements.txt
openalex>=0.5.0
openai>=1.0.0
apscheduler==3.10.4
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm>=4.66.0
